+++ changelog.md
@@ -8,6 +8,7 @@
     - Fixed cut/deleting of descriptions, was doing smallest to largest index, throwing off later #s, now reversed and does largest to smallest fixing the issue
     - Fixed cut/deleting descriptions was formatting terrain file wrong
     - Fixed shift click selection for room editor
+    - Fixed status bar better supported in virtual editor
 
 ## 0.4.53 2018-06-30